@@ -13,14 +13,12 @@
   "license": "ISC",
   "devDependencies": {
     "@babel/core": "^7.17.5",
-    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-env": "^7.26.9",
     "@babel/preset-react": "^7.16.7",
     "babel-loader": "^8.2.3",
-    "plotly.js": "^2.11.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-plotly.js": "^2.5.1",
-    "webpack": "^5.70.0",
-    "webpack-cli": "^4.9.2"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "webpack": "^5.98.0",
+    "webpack-cli": "^6.0.1"
   }
 }
